backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.0.350
chromadb==0.4.18
pypdf==3.17.1
pydub==0.25.1 